requirements.txt
fastapi
uvicorn
langchain
openai
openapi-schema-pydantic
networkx
gunicorn
python-multipart
PyPDF2
pypdf
chromadb
tiktoken
bs4
requests
typing
pinecone-client
